backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# LangChain RAG Stack (Constitution v1.1.0)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-qdrant>=0.2.0
langchain-text-splitters>=0.3.0

# Google Generative AI (for Gemini embeddings and LLM)
google-generativeai>=0.8.0

# Vector database - supports v1.7+ through v1.16+
qdrant-client>=1.12.0

# Data validation
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Environment management
python-dotenv>=1.0.0

# Utilities
httpx>=0.27.0
tiktoken>=0.8.0

# Database (optional - for session persistence)
asyncpg>=0.29.0
